move provider endpoints to the provider-api class
