+++ libs/common/src/admin-console/abstractions/provider/provider-api.service.abstraction.ts
@@ -1,6 +1,12 @@
+import { ProviderSetupRequest } from "../../models/request/provider/provider-setup.request";
+import { ProviderUpdateRequest } from "../../models/request/provider/provider-update.request";
 import { ProviderVerifyDeleteRecoverRequest } from "../../models/request/provider/provider-verify-delete-recover.request";
+import { ProviderResponse } from "../../models/response/provider/provider.response";
 
 export class ProviderApiServiceAbstraction {
+  postProviderSetup: (id: string, request: ProviderSetupRequest) => Promise<ProviderResponse>;
+  getProvider: (id: string) => Promise<ProviderResponse>;
+  putProvider: (id: string, request: ProviderUpdateRequest) => Promise<ProviderResponse>;
   providerRecoverDeleteToken: (
     organizationId: string,
     request: ProviderVerifyDeleteRecoverRequest,
